app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "jinja2>=3.1.6",
    "psycopg2-binary>=2.9.11",
    "python-multipart>=0.0.22",
    "uvicorn>=0.42.0",
]
